backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.9
alembic==1.10.3
psycopg2-binary==2.9.6
pydantic==1.10.7
google-auth==2.17.3
google-auth-oauthlib==1.0.0
email-validator==1.3.1
